# api/main.go
// Package main implements the defense-validation private ingress API (LLD §6.1).
//
// Step 1 scope: accept a SubmitDefenseValidationRequest@1 (LLD §10.1), validate it
// against the input contract, and return an accepted run reference (LLD §9.1).
// Downstream queueing, workers, and the verdict engine are out of scope for this step.
package main

import (
	"bytes"
	"context"
	_ "embed"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"crypto/rand"
)

const contractID = "defense-validation@1.0"

// runsPath is the durable async collection; compatRunsPath is the synchronous
// legacy collection. Both are declared here so the handlers strip exactly the
// prefix they are mounted on.
const (
	runsPath       = "/v1/defense-validation-runs"
	compatRunsPath = "/v1/compat/defense-validation-runs"
)

// upstreamContractID is the contract of the incoming defense-generation payload
// that now carries the mitigation rule (primary_candidate.artifact_content).
const upstreamContractID = "defense-generation@1.0"

//go:embed openapi.yaml
var openapiSpec []byte

// maxBodyBytes bounds request payloads at the API edge (LLD §3.4, §13.7).
const maxBodyBytes = 64 * 1024

// SubmitDefenseValidationRequest mirrors the request contract in LLD §10.1, plus
// inline artifact bodies. The reference IDs remain authoritative; the nested
// `substrate` / `candidate` / `test_basis` objects carry the actual artifact
// content (container image, WAF rule, attack test) so a run is self-describing.
type SubmitDefenseValidationRequest struct {
	ContractID          string `json:"contract_id"`
	RequestID           string `json:"request_id,omitempty"`
	CandidateArtifactID string `json:"candidate_artifact_id"`
	// TestBasisID is retained as an opaque producer selector for compatibility; no
	// test is derived from it, because nothing is executed here.
	TestBasisID    string          `json:"test_basis_id"`
	CheckProfileID string          `json:"check_profile_id"`
	Candidate      json.RawMessage `json:"candidate,omitempty"`
	// CorrelationID is echoed into the result envelope (optional).
	CorrelationID string `json:"correlation_id,omitempty"`
	// Reference-only mode is additive and mutually exclusive with inline artifacts
	// and legacy upstream_inputs. Both immutable producer locators are resolved and
	// verified only when this mode is selected.
	RoutePolicy   string                  `json:"route_policy,omitempty"`
	DefenseResult *ImmutableResultLocator `json:"defense_result,omitempty"`
	CheckResult   *ImmutableResultLocator `json:"check_result,omitempty"`
	// ProfileID selects an immutable route-adapter profile for the additive
	// shared-contract v2 path. Hydrated CG/DG bodies never enter the request.
	ProfileID string `json:"profile_id,omitempty"`

	// --- Upstream defense-generation payload (new input contract) ---
	// The mitigation rule is taken from primary_candidate.artifact_content. The
	// remaining envelope fields are accepted (so a full defense-generation result
	// validates) but not yet consumed; upstream_inputs will later carry the test.
	PrimaryCandidateRaw json.RawMessage `json:"primary_candidate,omitempty"`
	AttemptHistory      json.RawMessage `json:"attempt_history,omitempty"`
	OutcomeReason       json.RawMessage `json:"outcome_reason,omitempty"`
	ProofHandoffs       json.RawMessage `json:"proof_handoffs,omitempty"`
	UpstreamInputs      json.RawMessage `json:"upstream_inputs,omitempty"`
	UpstreamResultRef   json.RawMessage `json:"result_ref,omitempty"`
	ProducedAt          string          `json:"produced_at,omitempty"`
	UpstreamProse       string          `json:"prose_summary,omitempty"`
	UpstreamResultID    string          `json:"result_id,omitempty"`
	UpstreamTerminal    string          `json:"terminal_state,omitempty"`
	Callback            *CallbackSpec   `json:"callback,omitempty"`
}

type CallbackSpec struct {
	URL             string `json:"url"`
	EventContractID string `json:"event_contract_id"`
}

// PrimaryCandidate is the upstream defense-generation candidate. artifact_content
// carries the actual mitigation rule (a ModSecurity SecRule).
type PrimaryCandidate struct {
	ArtifactContent      string `json:"artifact_content"`
	ArtifactType         string `json:"artifact_type"`
	CandidateID          string `json:"candidate_id"`
	CandidateKind        string `json:"candidate_kind"`
	Discriminator        string `json:"discriminator"`
	SelectedControlClass string `json:"selected_control_class"`
}

// AcceptedRunResponse is the accepted run reference returned on submit (LLD §9.1).
type AcceptedRunResponse struct {
	RunID    string `json:"run_id"`
	ResultID string `json:"result_id"`
	Status   string `json:"status"`
}

// APIError is the controlled error envelope (LLD §9.5). It never carries stack
// traces, credentials, tokens, or unredacted attack payloads.
type APIError struct {
	Category string   `json:"category"`
	Message  string   `json:"message"`
	Fields   []string `json:"fields,omitempty"`
}

var store *RunStore
var dbx *DatabricksSink

// upstreamInputMode selects the separate upstream executor: the mitigation rule is
// read from a Databricks table referenced by the request's upstream_inputs, instead
// of the inline candidate. On by default; set env DV_INPUT_UPSTREAM=0 to disable.
var upstreamInputMode bool

// dbxReader reads upstream result_json rows from Databricks (upstream mode only).
var dbxReader *DatabricksReader

func main() {
	// CLI: print the custom WAF rule from a control-translation result.
	if len(os.Args) > 1 && os.Args[1] == "control-translation-waf-rule" {
		controlTranslationCLI(os.Args[2:])
		return
	}

	s, err := NewRunStore()
	if err != nil {
		log.Fatalf("could not open run ledger: %v", err)
	}
	store = s

	// Optional secondary sink (Databricks Delta). nil when DATABRICKS_DSN is unset.
	dbx = NewDatabricksSink()

	// Separate upstream executor (rule read from Databricks) — toggled by env, and
	// ON by default; set DV_INPUT_UPSTREAM to 0/false/no to use the legacy executor.
	v := strings.ToLower(strings.TrimSpace(os.Getenv("DV_INPUT_UPSTREAM")))
	if v == "" {
		v = "1"
	}
	if v == "1" || v == "true" || v == "yes" {
		upstreamInputMode = true
		dbxReader = NewDatabricksReader()
		log.Printf("input contract: UPSTREAM mode (rule read from Databricks via upstream_inputs)")
	}

	mux := http.NewServeMux()
	mux.HandleFunc(runsPath, withCORS(handleRunsCollection))
	mux.HandleFunc(runsPath+"/", withCORS(handleRunItem))
	mux.HandleFunc(compatRunsPath, withCORS(handleCompatibilityRunsCollection))
	mux.HandleFunc(compatRunsPath+"/", withCORS(handleCompatibilityRunItem))
	mux.HandleFunc("/healthz", withCORS(handleHealth))
	mux.HandleFunc("/openapi.yaml", withCORS(handleOpenAPI))
	mux.HandleFunc("/docs", withCORS(handleDocs))

	port := os.Getenv("PORT")
	if port == "" {
		port = "8190"
	}
	srv := &http.Server{Addr: ":" + port, Handler: mux}
	workerCtx, stopWorkers := context.WithCancel(context.Background())
	worker := NewRunWorker(store, executeDurableRun, dbx)
	callbackDispatcher := NewCallbackDispatcher(store, os.Getenv("CAPABILITY_CALLBACK_TOKEN"))
	go worker.Run(workerCtx)
	go callbackDispatcher.Run(workerCtx)

	// On SIGINT/SIGTERM (docker stop) stop accepting new requests and drain
	// in-flight ones, then fall through to close the db pool cleanly.
	go func() {
		sig := make(chan os.Signal, 1)
		signal.Notify(sig, syscall.SIGINT, syscall.SIGTERM)
		<-sig
		log.Print("shutting down…")
		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
		defer cancel()
		_ = srv.Shutdown(ctx)
	}()

	log.Printf("defense-validation API listening on %s", srv.Addr)
	err = srv.ListenAndServe()
	stopWorkers()
	store.Close() // close the connection pool after in-flight requests drain
	dbx.Close()
	dbxReader.Close()
	if err != nil && err != http.ErrServerClosed {
		log.Fatal(err)
	}
}

// handleOpenAPI serves the embedded OpenAPI 3 spec.
// enrichEnvelope fills the result-envelope fields on the outcome: capability /
// contract, workflow status, the echoed correlation id, evidence refs, and a
// result_ref pointing at the Databricks row (keyed by run_id + result_id, the
// same values written to that table so a consumer can query it).
func enrichEnvelope(out *RunOutcome, correlationID string) {
	out.Capability = "defense-validation"
	out.ContractID = contractID
	out.CorrelationID = correlationID
	if out.EvidenceRefs == nil {
		out.EvidenceRefs = []string{}
	}
	if out.TerminalState == stateMalfunction {
		out.Status = "failed"
	} else {
		out.Status = "completed"
	}
	out.ResultRef = nil
	if dbx != nil {
		out.ResultRef = dbx.ResultRef(out.ResultID)
	}
}

func handleOpenAPI(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/yaml")
	_, _ = w.Write(openapiSpec)
}

// handleDocs serves a Swagger UI page pointed at the embedded spec. The Swagger
// UI assets load from a CDN, so viewing /docs needs internet access.
func handleDocs(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	_, _ = w.Write([]byte(swaggerHTML))
}

const swaggerHTML = `<!doctype html>
<html lang="en">
<head>
  <meta charset="utf-8"/>
  <meta name="viewport" content="width=device-width, initial-scale=1"/>
  <title>defense-validation API — Swagger</title>
  <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css"/>
</head>
<body>
  <div id="swagger-ui"></div>
  <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
  <script>
    window.ui = SwaggerUIBundle({
      url: 'openapi.yaml',
      dom_id: '#swagger-ui',
      deepLinking: true,
      tryItOutEnabled: true
    });
  </script>
</body>
</html>`

func handleHealth(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		writeError(w, http.StatusMethodNotAllowed, APIError{Category: "invalid-input", Message: "method not allowed"})
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

// handleCompatibilityRunsCollection retains the old synchronous API on an
// explicit compatibility path. New orchestrators must use the durable async API.
func handleCompatibilityRunsCollection(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		writeJSON(w, http.StatusOK, store.List())
	case http.MethodPost:
		handleSubmitRunSync(w, r)
	default:
		writeError(w, http.StatusMethodNotAllowed, APIError{
			Category: "invalid-input", Message: "method not allowed",
		})
	}
}

// handleCompatibilityRunItem returns one legacy run's immutable request + response.
func handleCompatibilityRunItem(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		writeError(w, http.StatusMethodNotAllowed, APIError{
			Category: "invalid-input", Message: "method not allowed",
		})
		return
	}
	id := runIDFromPath(r.URL.Path, compatRunsPath)
	rec, ok := store.Get(id)
	if !ok {
		writeError(w, http.StatusNotFound, APIError{
			Category: "invalid-input", Message: "run not found",
		})
		return
	}
	writeJSON(w, http.StatusOK, rec)
}

// handleSubmitRunSync implements the explicitly separate synchronous compatibility path.
func handleSubmitRunSync(w http.ResponseWriter, r *http.Request) {
	req, raw, apiErr := decodeRequest(r)
	if apiErr != nil {
		writeError(w, http.StatusBadRequest, *apiErr)
		return
	}

	if fields := validate(req); len(fields) > 0 {
		writeError(w, http.StatusBadRequest, APIError{
			Category: "invalid-input",
			Message:  "request does not satisfy defense-validation@1.0 contract",
			Fields:   fields,
		})
		return
	}

	// Execute the scenario synchronously: bring up the substrate container, apply
	// the candidate WAF rule, run the supplied test, and resolve a terminal state
	// (LLD §5, §6.4, §6.5). The request context always remains authoritative;
	// DV_EXECUTION_TIMEOUT can optionally add an operator-configured deadline.
	runID := "dv-run-" + newID()
	resultID := resultIDPrefix + newID()

	ctx, cancel := executionContext(r.Context())
	defer cancel()

	outcome := resolveRequestedRule(ctx, req, runID, resultID)
	enrichEnvelope(&outcome, req.CorrelationID)
	outcome.CreatedAt = time.Now().UTC()
	if err := setCanonicalIntegrity(&outcome); err != nil {
		writeError(w, http.StatusInternalServerError, APIError{Category: "malfunction", Message: "could not encode canonical result"})
		return
	}
	payload, err := canonicalResultPayload(outcome)
	if err != nil {
		writeError(w, http.StatusInternalServerError, APIError{Category: "malfunction", Message: "could not encode canonical result"})
		return
	}

	// Secondary sink: write the result to the Databricks table synchronously so the
	// envelope's status can reflect the write. terminal_state stays the test result;
	// only status degrades to "storage-failed" when the write fails (the run is not
	// a hard failure). A malfunction is already "failed", so a write failure there
	// doesn't change it — but the row is still written for diagnostics.
	if dbx != nil {
		if publishErr := dbx.Publish(ctx, outcome, payload); publishErr != nil {
			verifyErr := dbx.Verify(context.WithoutCancel(ctx), outcome, payload)
			if verifyErr != nil && outcome.TerminalState != stateMalfunction {
				outcome.Status = "storage-failed"
				if err := setCanonicalIntegrity(&outcome); err == nil {
					payload, _ = canonicalResultPayload(outcome)
				}
			}
		}
	}

	// Record the run in the ledger (source of truth) with the exact immutable
	// request bytes and the final envelope (including the resolved status).
	if err := store.Add(&RunRecord{
		RunID:         runID,
		ResultID:      resultID,
		TerminalState: outcome.TerminalState,
		CreatedAt:     time.Now().UTC(),
		Request:       raw,
		Response:      outcome,
	}); err != nil {
		log.Printf("ledger: failed to persist run %s: %v", runID, err)
	}

	writeJSON(w, http.StatusOK, outcome)
}

// decodeRequest reads and strictly decodes the body, rejecting unknown fields to
// honor the contract's additionalProperties:false (LLD §10.1). It also returns the
// exact raw bytes so the run ledger can store the immutable request.
func decodeRequest(r *http.Request) (SubmitDefenseValidationRequest, json.RawMessage, *APIError) {
	var req SubmitDefenseValidationRequest

	raw, err := io.ReadAll(http.MaxBytesReader(nil, r.Body, maxBodyBytes))
	if err != nil {
		var maxErr *http.MaxBytesError
		if errors.As(err, &maxErr) {
			return req, nil, &APIError{Category: "invalid-input", Message: "request body too large"}
		}
		return req, nil, &APIError{Category: "invalid-input", Message: "could not read request body"}
	}
	if len(raw) == 0 {
		return req, nil, &APIError{Category: "invalid-input", Message: "request body is empty"}
	}

	dec := json.NewDecoder(bytes.NewReader(raw))
	dec.DisallowUnknownFields()
	if err := dec.Decode(&req); err != nil {
		if strings.Contains(err.Error(), "unknown field") {
			return req, nil, &APIError{
				Category: "invalid-input",
				Message:  "request contains fields not permitted by the contract",
			}
		}
		return req, nil, &APIError{Category: "invalid-input", Message: "request body is not valid JSON"}
	}

	// Guard against trailing garbage after the JSON object.
	if err := dec.Decode(&struct{}{}); err != io.EOF {
		return req, nil, &APIError{Category: "invalid-input", Message: "request body must contain a single JSON object"}
	}
	return req, json.RawMessage(raw), nil
}

// validate enforces the SubmitDefenseValidationRequest@1 field rules (LLD §10.1)
// and returns the names of every offending field.
func validate(req SubmitDefenseValidationRequest) []string {
	var bad []string
	// In upstream-input mode the rule is read from Databricks via upstream_inputs, so
	// the legacy reference ids are not required and the upstream contract is accepted.
	if req.ContractID != contractID {
		bad = append(bad, "contract_id")
	}
	if v2LocatorMode(req) {
		return append(bad, validateV2LocatorRequest(req)...)
	}
	if locatorMode(req) {
		return append(bad, validateLocatorRequest(req)...)
	}
	if strings.TrimSpace(req.CandidateArtifactID) == "" {
		bad = append(bad, "candidate_artifact_id")
	}
	if strings.TrimSpace(req.TestBasisID) == "" {
		bad = append(bad, "test_basis_id")
	}
	if strings.TrimSpace(req.CheckProfileID) == "" {
		bad = append(bad, "check_profile_id")
	}
	if len(req.UpstreamInputs) > 0 {
		entries, err := parseUpstreamInputs(req.UpstreamInputs)
		if err != nil {
			bad = append(bad, "upstream_inputs")
		} else {
			for i, entry := range entries {
				prefix := fmt.Sprintf("upstream_inputs[%d]", i)
				if strings.TrimSpace(entry.Capability) == "" || strings.TrimSpace(entry.ContractID) == "" {
					bad = append(bad, prefix)
				}
				ref := entry.ResultRef
				if ref.System != "databricks" || strings.TrimSpace(ref.Catalog) == "" || strings.TrimSpace(ref.Schema) == "" ||
					strings.TrimSpace(ref.Table) == "" || strings.TrimSpace(ref.Key) == "" || ref.Key != entry.ResultID {
					bad = append(bad, prefix+".result_ref")
				}
			}
		}
	}

	// The candidate body is optional, but validated when present: it is the rule
	// this capability reports.
	if len(req.Candidate) > 0 {
		var c CandidateSpec
		if err := json.Unmarshal(req.Candidate, &c); err != nil {
			bad = append(bad, "candidate")
		} else {
			if c.Kind == "" {
				bad = append(bad, "candidate.kind")
			}
			if strings.TrimSpace(c.Rule) == "" {
				bad = append(bad, "candidate.rule")
			}
		}
	}
	return bad
}

// runIDFromPath returns the path remainder after the collection a handler is
// mounted on, or "" when path is not under base. A handler must pass its own
// base: the durable and compatibility collections share a suffix, so a bare
// TrimPrefix of the wrong one is a silent no-op that leaves the whole path in
// place and makes every ledger lookup miss. Returning "" on a non-match turns
// that mistake into an immediate not-found instead.
func runIDFromPath(path, base string) string {
	rest, ok := strings.CutPrefix(path, base+"/")
	if !ok {
		return ""
	}
	return strings.Trim(rest, "/")
}

func withCORS(next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Access-Control-Allow-Origin", "*")
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Idempotency-Key, X-Correlation-ID")
		if r.Method == http.MethodOptions {
			w.WriteHeader(http.StatusNoContent)
			return
		}
		next(w, r)
	}
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}

func writeError(w http.ResponseWriter, status int, e APIError) {
	writeJSON(w, status, map[string]APIError{"error": e})
}

// newID returns a short random hex identifier for run/result references.
func newID() string {
	var b [12]byte
	if _, err := rand.Read(b[:]); err != nil {
		return fmt.Sprintf("%d", time.Now().UnixNano())
	}
	return fmt.Sprintf("%x", b)
}