# api/openapi.yaml
openapi: 3.0.3
info:
  title: defense-validation API
  version: "1.0"
  description: >
    Resolves one defensive control candidate from a verified producer result and
    reports it (contract `defense-validation@1.0`). The rule may come from a
    defense-generation or control-translation row named by `upstream_inputs`, from
    immutable Check Generation and Defense Generation locators, or inline in the
    request. The `shared-attack-contracts-v2` route additionally reads back the
    complete candidate bundle's application unit.

    This capability does NOT judge the rule: nothing is executed, no traffic is
    sent, and the result carries no verdict — no expected/actual, no match, no
    substrate. Pushing the rule to a third-party control plane is the next stage,
    and any pass/fail determination belongs to that plane.

    Canonical submission is asynchronous and creates a durable PostgreSQL run
    before a leased worker resolves the rule. Valid transitions are queued to
    running to a terminal state, queued to canceled, running to queued for
    publication retry, and expired running to failed or canceled during recovery.
    Terminal runs are immutable.
servers:
  - url: http://localhost:8137
    description: Local / Docker Compose
paths:
  /v1/defense-validation-runs:
    post:
      summary: Submit a durable asynchronous defense-validation run
      operationId: submitRun
      description: >
        Validates and durably queues work, then immediately returns. Identical
        retries reuse the run; reuse with a different normalized digest is 409.
        Only the application/json media type is accepted. The three optional
        X-Janus-Callback-* headers enable a durable terminal-state wakeup;
        polling remains the canonical fallback.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
        - $ref: "#/components/parameters/CallbackURL"
        - $ref: "#/components/parameters/CallbackWorkflowID"
        - $ref: "#/components/parameters/CallbackSignal"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDefenseValidationRequest"
            examples:
              inlineRule:
                summary: Report a rule supplied inline
                value:
                  contract_id: defense-validation@1.0
                  candidate_artifact_id: candidate:log4shell:waf-rule:strict
                  test_basis_id: test-basis:log4shell:attack:jndi-ldap
                  check_profile_id: defense-validation-profile:waf-http:1
                  candidate:
                    kind: waf-rule
                    engine: modsecurity
                    rule_id: "1005440"
                    action: deny
                    rule: >-
                      SecRule REQUEST_HEADERS|REQUEST_URI|ARGS "@rx (?i)\$\{jndi:(?:ldaps?|rmi|dns):/[^}]*\}" "id:1005440,phase:2,deny,status:403"
              upstreamRow:
                summary: Resolve the rule from a control-translation row
                value:
                  contract_id: defense-validation@1.0
                  request_id: janus-defense-validation-request-1
                  correlation_id: correlation-1
                  candidate_artifact_id: candidate:akamai-waf:e6fcf665
                  test_basis_id: unused-in-this-mode
                  check_profile_id: defense-validation-profile:waf-http:1
                  upstream_inputs:
                    - capability: control-translation
                      contract_id: control-translation-result@2.0
                      result_id: control-translation-result:e0b7e5cf
                      result_ref:
                        system: databricks
                        catalog: 36889_janus_dev
                        schema: control_translation
                        table: control_translation_results
                        key: control-translation-result:e0b7e5cf
      responses:
        "202":
          description: Run durably queued or running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "200":
          description: Idempotent replay of an already-terminal run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          description: Request does not satisfy the contract.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          description: Content-Type is not application/json.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: The idempotency key is bound to another request digest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      summary: List runs (newest first)
      operationId: listRuns
      responses:
        "200":
          description: Run summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunSummary"
...
  /v1/compat/defense-validation-runs:
    post:
      summary: Execute synchronously for legacy UI compatibility
      deprecated: true
      operationId: submitRunSynchronously
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDefenseValidationRequest"
      responses:
        "200":
          description: Full terminal result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunOutcome"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1 }
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: true
      schema: { type: string, minLength: 1 }
    CallbackURL:
      name: X-Janus-Callback-URL
      in: header
      required: false
      description: HTTPS orchestration callback endpoint. Required with the other callback headers.
      schema: { type: string, format: uri, pattern: "^https://" }
    CallbackWorkflowID:
      name: X-Janus-Callback-Workflow-ID
      in: header
      required: false
      description: Temporal child workflow ID only; a workflow run ID is never accepted or required.
      schema: { type: string, minLength: 1 }
    CallbackSignal:
      name: X-Janus-Callback-Signal
      in: header
      required: false
      description: Required with the other callback headers.
      schema: { type: string, enum: ["janus.capability-completion.v1"] }
...
  schemas:
    SubmitDefenseValidationRequest:
      type: object
      additionalProperties: false
      required: [contract_id]
      anyOf:
        - required: [candidate_artifact_id, test_basis_id, check_profile_id]
        - required: [request_id, correlation_id, route_policy, defense_result, check_result]
        - required: [request_id, correlation_id, route_policy, defense_result, check_result, profile_id]
      properties:
        contract_id:
          type: string
          enum: ["defense-validation@1.0"]
        request_id:
          type: string
          minLength: 1
          description: Must equal Idempotency-Key. May be omitted during migration and is then populated from that header.
        candidate_artifact_id:
          type: string
          minLength: 1
          description: Immutable candidate mitigation reference.
        test_basis_id:
          type: string
          minLength: 1
          description: >
            Opaque producer selector, retained for compatibility. No test is
            derived from it: nothing is executed here.
        check_profile_id:
          type: string
          minLength: 1
          description: Signed check profile reference.
        correlation_id:
          type: string
          description: >
            Must equal X-Correlation-ID. May be omitted during migration and is
            then populated from that required header.
          example: "mc-request:CVE-2021-44228:waf:1"
        route_policy:
          type: string
          enum: [registered-waf-route-v1, shared-attack-contracts-v2]
          description: >
            Selects strict reference-only hydration. registered-waf-route-v1 is
            the legacy single-case route. shared-attack-contracts-v2 verifies the
            complete CG semantics and DG application unit, then executes every
            exact obligation/input assignment. Neither route accepts executable
            candidate or test bodies.
        profile_id:
          type: string
          enum: [waf-standard@1, waf-standard@2]
          description: >
            Required only for shared-attack-contracts-v2. Selects the embedded,
            immutable MC route-adapter profile used for destination-free HTTP templates.
        defense_result:
          allOf:
            - $ref: "#/components/schemas/ImmutableResultLocator"
          description: Immutable Defense Generation result locator. Required only in reference-only mode.
        check_result:
          allOf:
            - $ref: "#/components/schemas/ImmutableResultLocator"
          description: Immutable Check Generation result locator. Required only in reference-only mode.
        candidate:
          allOf:
            - $ref: "#/components/schemas/CandidateSpec"
          description: >
            The rule, supplied inline. Accepted only in inline mode: the locator
            and shared-contract routes reject caller-supplied bodies. Inline mode
            carries no producer lineage, so the rule is not verified against any
            upstream result.
        upstream_inputs:
          type: array
          description: >
            Upstream-mode input (env DV_INPUT_UPSTREAM). Each entry's result_ref
            points at a Databricks row. The rule is read from the
            defense-generation entry's result_json.primary_candidate.artifact_content,
            or, when there is no defense-generation entry, from the
            control-translation entry's artifacts map (resolved by artifact_id and
            verified against its content_hash). The top-level contract remains
            "defense-validation@1.0" and canonical identity fields remain required.
          items:
            $ref: "#/components/schemas/UpstreamInput"
        primary_candidate:
          type: object
          description: >
            Upstream defense-generation candidate; artifact_content carries the
            ModSecurity rule. Accepted so a full defense-generation payload
            validates. In upstream mode the rule is read from Databricks, not here.
        callback:
          type: object
          deprecated: true
          description: >
            Unsupported body-level callback metadata. Use the three
            X-Janus-Callback-* request headers instead.
          additionalProperties: false
          required: [url]
          properties:
            url: { type: string, format: uri, pattern: "^https://" }
            event_contract_id: { type: string, enum: ["capability-run-event@1.0"] }
    SubmissionResponse:
      type: object
      required: [capability, contract_id, request_id, correlation_id, run_id, status, status_url, result_url, accepted_at]
      properties:
        capability: { type: string, enum: [defense-validation] }
        contract_id: { type: string, enum: ["defense-validation-run-submission@1.0"] }
        request_id: { type: string }
        correlation_id: { type: string }
        run_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed, canceled] }
        status_url: { type: string }
        result_url: { type: string }
        accepted_at: { type: string, format: date-time }
...
    RunOutcome:
      type: object
      description: >
        Result envelope plus the resolved rule and how it was obtained. There is
        no verdict here: no expected/actual, no match, no substrate.
      required: [capability, contract_id, request_id, correlation_id, run_id, result_id, terminal_state, status, result_ref, evidence_refs, request_sha256, prose_summary, content_sha256, size_bytes, created_at]
      properties:
        capability:
          type: string
          enum: ["defense-validation"]
        contract_id:
          type: string
          enum: ["defense-validation@1.0"]
        request_id:
          type: string
        run_id:
          type: string
        result_id:
          type: string
        terminal_state:
          type: string
          enum: [rule-resolved, failed, malfunction]
          description: >
            rule-resolved means a rule was read from a verified producer result and
            reported; it says nothing about whether the rule is effective. failed
            means no rule could be resolved. malfunction is an internal fault.
        status:
          type: string
          enum: [completed, failed]
          description: >
            Lifecycle status. Canonical asynchronous completion is reported only
            after immutable Databricks publication succeeds.
        correlation_id:
          type: string
          description: Echoed from the request; omitted when not supplied.
        result_ref:
          $ref: "#/components/schemas/ResultRef"
        evidence_refs:
          type: array
          items:
            type: string
        request_sha256:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
          description: SHA-256 of the normalized request bound to the idempotency key.
        upstream_inputs:
          type: array
          description: Exact authoritative upstream result identities the rule was resolved from.
          items:
            $ref: "#/components/schemas/UpstreamInput"
        input_provenance:
          type: object
          additionalProperties: false
          required: [route_policy, defense_result, check_result, verification]
          properties:
            route_policy: { type: string, enum: [registered-waf-route-v1, shared-attack-contracts-v2] }
            defense_result: { $ref: "#/components/schemas/ImmutableResultLocator" }
            check_result: { $ref: "#/components/schemas/ImmutableResultLocator" }
            verification: { type: string, enum: [physical-and-logical-sha256-verified] }
        profile_id:
          type: string
          enum: [waf-standard@1, waf-standard@2]
          description: Present only for shared-contract v2 results.
        application_unit:
          allOf:
            - $ref: "#/components/schemas/AppliedApplicationUnit"
          description: >
            Which artifact set was read back for the shared-contract route. This is
            provenance for the rule, not a judgement about it.
        candidate:
          allOf:
            - $ref: "#/components/schemas/CandidateSpec"
          description: >
            The resolved rule (embedded so a defense_validation row is
            self-contained; consumers need not query the upstream table). Stripped
            from the API result response, present in the canonical row.
        detail:
          type: string
          description: Why a run failed, or how the rule was obtained on success.
        steps:
          type: array
          items:
            type: string
        prose_summary:
          type: string
        limitations:
          type: array
          items:
            type: string
        content_sha256:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        size_bytes:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
...
    RunSummary:
      type: object
      properties:
        run_id:
          type: string
        result_id:
          type: string
        terminal_state:
          type: string
        created_at:
          type: string
          format: date-time
        summary:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [code, detail, retryable]
      properties:
        code:
          type: string
          example: invalid_request
        detail:
          type: string
        retryable:
          type: boolean